+++ src/detail/llvm_helpers.cpp
@@ -24,7 +24,6 @@
 #include <vector>
 
 #include <boost/math/constants/constants.hpp>
-#include <boost/math/special_functions/binomial.hpp>
 #include <boost/numeric/conversion/cast.hpp>
 
 #include <fmt/format.h>
@@ -54,8 +53,10 @@
 
 #endif
 
+#include <heyoka/detail/binomial.hpp>
 #include <heyoka/detail/llvm_fwd.hpp>
 #include <heyoka/detail/llvm_helpers.hpp>
+#include <heyoka/detail/llvm_vector_type.hpp>
 #include <heyoka/detail/logging_impl.hpp>
 #include <heyoka/detail/sleef.hpp>
 #include <heyoka/detail/visibility.hpp>
@@ -93,19 +94,28 @@ const auto type_map = []() {
         };
     }
 
-    // Try to associate C++ long double to LLVM double or x86_fp80.
+    // Try to associate C++ long double to an LLVM fp type.
     if (std::numeric_limits<long double>::is_iec559) {
         if (std::numeric_limits<long double>::digits == 53) {
             retval[typeid(long double)] = [](llvm::LLVMContext &c) {
-                // IEEE double-precision type (this is the case on MSVC for instance).
+                // IEEE double-precision format (this is the case on MSVC for instance).
                 auto ret = llvm::Type::getDoubleTy(c);
                 assert(ret != nullptr);
                 return ret;
             };
+#if defined(HEYOKA_ARCH_X86)
         } else if (std::numeric_limits<long double>::digits == 64) {
             retval[typeid(long double)] = [](llvm::LLVMContext &c) {
                 // x86 extended precision format.
                 auto ret = llvm::Type::getX86_FP80Ty(c);
+                assert(ret != nullptr);
+                return ret;
+            };
+#endif
+        } else if (std::numeric_limits<long double>::digits == 113) {
+            retval[typeid(long double)] = [](llvm::LLVMContext &c) {
+                // IEEE quadruple-precision format (e.g., ARM 64).
+                auto ret = llvm::Type::getFP128Ty(c);
                 assert(ret != nullptr);
                 return ret;
             };
@@ -146,7 +156,7 @@ std::string llvm_mangle_type(llvm::Type *t)
 {
     assert(t != nullptr);
 
-    if (auto *v_t = llvm::dyn_cast<llvm::VectorType>(t)) {
+    if (auto *v_t = llvm::dyn_cast<llvm_vector_type>(t)) {
         // If the type is a vector, get the name of the element type
         // and append the vector size.
         return "{}_{}"_format(llvm_type_name(v_t->getElementType()), v_t->getNumElements());
@@ -191,7 +201,7 @@ llvm::Value *load_vector_from_memory(ir_builder &builder, llvm::Value *ptr, std:
 // a plain store will be performed.
 void store_vector_to_memory(ir_builder &builder, llvm::Value *ptr, llvm::Value *vec)
 {
-    if (auto v_ptr_t = llvm::dyn_cast<llvm::VectorType>(vec->getType())) {
+    if (auto v_ptr_t = llvm::dyn_cast<llvm_vector_type>(vec->getType())) {
         // Determine the vector size.
         const auto vector_size = boost::numeric_cast<std::uint32_t>(v_ptr_t->getNumElements());
 
@@ -236,13 +246,7 @@ llvm::Type *make_vector_type(llvm::Type *t, std::uint32_t vector_size)
     if (vector_size == 1u) {
         return t;
     } else {
-        auto retval =
-#if LLVM_VERSION_MAJOR == 10
-            llvm::VectorType::get
-#else
-            llvm::FixedVectorType::get
-#endif
-            (t, boost::numeric_cast<unsigned>(vector_size));
+        auto retval = llvm_vector_type::get(t, boost::numeric_cast<unsigned>(vector_size));
 
         assert(retval != nullptr);
 
@@ -254,7 +258,7 @@ llvm::Type *make_vector_type(llvm::Type *t, std::uint32_t vector_size)
 // return {vec}.
 std::vector<llvm::Value *> vector_to_scalars(ir_builder &builder, llvm::Value *vec)
 {
-    if (auto vec_t = llvm::dyn_cast<llvm::VectorType>(vec->getType())) {
+    if (auto vec_t = llvm::dyn_cast<llvm_vector_type>(vec->getType())) {
         // Fetch the vector width.
         auto vector_size = vec_t->getNumElements();
 
@@ -836,7 +840,7 @@ std::pair<llvm::Value *, llvm::Value *> llvm_sincos(llvm_state &s, llvm::Value *
         return {scalars_to_vector(builder, res_sin), scalars_to_vector(builder, res_cos)};
     } else {
 #endif
-        if (auto vec_t = llvm::dyn_cast<llvm::VectorType>(x->getType())) {
+        if (auto vec_t = llvm::dyn_cast<llvm_vector_type>(x->getType())) {
             // NOTE: although there exists a SLEEF function for computing sin/cos
             // at the same time, we cannot use it directly because it returns a pair
             // of SIMD vectors rather than a single one and that does not play
@@ -1063,7 +1067,7 @@ llvm::Value *llvm_sgn(llvm_state &s, llvm::Value *val)
 
     // Convert to int32.
     llvm::Type *int_type;
-    if (auto *v_t = llvm::dyn_cast<llvm::VectorType>(cmp0->getType())) {
+    if (auto *v_t = llvm::dyn_cast<llvm_vector_type>(cmp0->getType())) {
         int_type = make_vector_type(builder.getInt32Ty(), boost::numeric_cast<std::uint32_t>(v_t->getNumElements()));
     } else {
         int_type = builder.getInt32Ty();
@@ -1539,9 +1543,7 @@ llvm::Value *llvm_add_bc_array_impl(llvm_state &s, std::uint32_t n)
         for (std::uint32_t j = 0; j <= n; ++j) {
             // NOTE: the Boost implementation requires j <= i. We don't care about
             // j > i anyway.
-            const auto val = (j <= i) ? boost::math::binomial_coefficient<T>(boost::numeric_cast<unsigned>(i),
-                                                                             boost::numeric_cast<unsigned>(j))
-                                      : T(0);
+            const auto val = (j <= i) ? binomial<T>(i, j) : T(0);
             bc_const.push_back(llvm::cast<llvm::Constant>(codegen<T>(s, number{val})));
         }
     }