+++ src/detail/sleef.cpp
@@ -21,8 +21,6 @@
 #include <llvm/IR/LLVMContext.h>
 #include <llvm/IR/Type.h>
 
-#include <sleef.h>
-
 #include <heyoka/detail/llvm_fwd.hpp>
 #include <heyoka/detail/sleef.hpp>
 #include <heyoka/detail/visibility.hpp>
@@ -54,6 +52,9 @@ struct sleef_key_hasher {
 using sleef_map_t = std::unordered_map<sleef_key_t, std::string, sleef_key_hasher>;
 
 // Helper to construct the sleef map for the double-precision type.
+// NOTE: at this time the sleef conda package for PPC64 does not seem
+// to provide VSX3 functions. Thus, for now we use only the
+// VSX implementations.
 auto make_sleef_map_dbl()
 {
     const auto &features = get_target_features();
@@ -73,6 +74,10 @@ auto make_sleef_map_dbl()
         retval[{"sin", 2}] = "Sleef_sind2_u10sse4";
     } else if (features.sse2) {
         retval[{"sin", 2}] = "Sleef_sind2_u10sse2";
+    } else if (features.aarch64) {
+        retval[{"sin", 2}] = "Sleef_sind2_u10advsimd";
+    } else if (features.vsx) {
+        retval[{"sin", 2}] = "Sleef_sind2_u10vsx";
     }
 
     // cos().
@@ -88,6 +93,10 @@ auto make_sleef_map_dbl()
         retval[{"cos", 2}] = "Sleef_cosd2_u10sse4";
     } else if (features.sse2) {
         retval[{"cos", 2}] = "Sleef_cosd2_u10sse2";
+    } else if (features.aarch64) {
+        retval[{"cos", 2}] = "Sleef_cosd2_u10advsimd";
+    } else if (features.vsx) {
+        retval[{"cos", 2}] = "Sleef_cosd2_u10vsx";
     }
 
     // log().
@@ -103,6 +112,10 @@ auto make_sleef_map_dbl()
         retval[{"log", 2}] = "Sleef_logd2_u10sse4";
     } else if (features.sse2) {
         retval[{"log", 2}] = "Sleef_logd2_u10sse2";
+    } else if (features.aarch64) {
+        retval[{"log", 2}] = "Sleef_logd2_u10advsimd";
+    } else if (features.vsx) {
+        retval[{"log", 2}] = "Sleef_logd2_u10vsx";
     }
 
     // exp().
@@ -118,6 +131,10 @@ auto make_sleef_map_dbl()
         retval[{"exp", 2}] = "Sleef_expd2_u10sse4";
     } else if (features.sse2) {
         retval[{"exp", 2}] = "Sleef_expd2_u10sse2";
+    } else if (features.aarch64) {
+        retval[{"exp", 2}] = "Sleef_expd2_u10advsimd";
+    } else if (features.vsx) {
+        retval[{"exp", 2}] = "Sleef_expd2_u10vsx";
     }
 
     // pow().
@@ -133,6 +150,10 @@ auto make_sleef_map_dbl()
         retval[{"pow", 2}] = "Sleef_powd2_u10sse4";
     } else if (features.sse2) {
         retval[{"pow", 2}] = "Sleef_powd2_u10sse2";
+    } else if (features.aarch64) {
+        retval[{"pow", 2}] = "Sleef_powd2_u10advsimd";
+    } else if (features.vsx) {
+        retval[{"pow", 2}] = "Sleef_powd2_u10vsx";
     }
 
     // tan().
@@ -148,6 +169,10 @@ auto make_sleef_map_dbl()
         retval[{"tan", 2}] = "Sleef_tand2_u10sse4";
     } else if (features.sse2) {
         retval[{"tan", 2}] = "Sleef_tand2_u10sse2";
+    } else if (features.aarch64) {
+        retval[{"tan", 2}] = "Sleef_tand2_u10advsimd";
+    } else if (features.vsx) {
+        retval[{"tan", 2}] = "Sleef_tand2_u10vsx";
     }
 
     // asin().
@@ -163,6 +188,10 @@ auto make_sleef_map_dbl()
         retval[{"asin", 2}] = "Sleef_asind2_u10sse4";
     } else if (features.sse2) {
         retval[{"asin", 2}] = "Sleef_asind2_u10sse2";
+    } else if (features.aarch64) {
+        retval[{"asin", 2}] = "Sleef_asind2_u10advsimd";
+    } else if (features.vsx) {
+        retval[{"asin", 2}] = "Sleef_asind2_u10vsx";
     }
 
     // acos().
@@ -178,6 +207,10 @@ auto make_sleef_map_dbl()
         retval[{"acos", 2}] = "Sleef_acosd2_u10sse4";
     } else if (features.sse2) {
         retval[{"acos", 2}] = "Sleef_acosd2_u10sse2";
+    } else if (features.aarch64) {
+        retval[{"acos", 2}] = "Sleef_acosd2_u10advsimd";
+    } else if (features.vsx) {
+        retval[{"acos", 2}] = "Sleef_acosd2_u10vsx";
     }
 
     // atan().
@@ -193,6 +226,10 @@ auto make_sleef_map_dbl()
         retval[{"atan", 2}] = "Sleef_atand2_u10sse4";
     } else if (features.sse2) {
         retval[{"atan", 2}] = "Sleef_atand2_u10sse2";
+    } else if (features.aarch64) {
+        retval[{"atan", 2}] = "Sleef_atand2_u10advsimd";
+    } else if (features.vsx) {
+        retval[{"atan", 2}] = "Sleef_atand2_u10vsx";
     }
 
     // cosh().
@@ -208,6 +245,10 @@ auto make_sleef_map_dbl()
         retval[{"cosh", 2}] = "Sleef_coshd2_u10sse4";
     } else if (features.sse2) {
         retval[{"cosh", 2}] = "Sleef_coshd2_u10sse2";
+    } else if (features.aarch64) {
+        retval[{"cosh", 2}] = "Sleef_coshd2_u10advsimd";
+    } else if (features.vsx) {
+        retval[{"cosh", 2}] = "Sleef_coshd2_u10vsx";
     }
 
     // sinh().
@@ -223,6 +264,10 @@ auto make_sleef_map_dbl()
         retval[{"sinh", 2}] = "Sleef_sinhd2_u10sse4";
     } else if (features.sse2) {
         retval[{"sinh", 2}] = "Sleef_sinhd2_u10sse2";
+    } else if (features.aarch64) {
+        retval[{"sinh", 2}] = "Sleef_sinhd2_u10advsimd";
+    } else if (features.vsx) {
+        retval[{"sinh", 2}] = "Sleef_sinhd2_u10vsx";
     }
 
     // tanh().
@@ -238,6 +283,10 @@ auto make_sleef_map_dbl()
         retval[{"tanh", 2}] = "Sleef_tanhd2_u10sse4";
     } else if (features.sse2) {
         retval[{"tanh", 2}] = "Sleef_tanhd2_u10sse2";
+    } else if (features.aarch64) {
+        retval[{"tanh", 2}] = "Sleef_tanhd2_u10advsimd";
+    } else if (features.vsx) {
+        retval[{"tanh", 2}] = "Sleef_tanhd2_u10vsx";
     }
 
     // asinh().
@@ -253,6 +302,10 @@ auto make_sleef_map_dbl()
         retval[{"asinh", 2}] = "Sleef_asinhd2_u10sse4";
     } else if (features.sse2) {
         retval[{"asinh", 2}] = "Sleef_asinhd2_u10sse2";
+    } else if (features.aarch64) {
+        retval[{"asinh", 2}] = "Sleef_asinhd2_u10advsimd";
+    } else if (features.vsx) {
+        retval[{"asinh", 2}] = "Sleef_asinhd2_u10vsx";
     }
 
     // acosh().
@@ -268,6 +321,10 @@ auto make_sleef_map_dbl()
         retval[{"acosh", 2}] = "Sleef_acoshd2_u10sse4";
     } else if (features.sse2) {
         retval[{"acosh", 2}] = "Sleef_acoshd2_u10sse2";
+    } else if (features.aarch64) {
+        retval[{"acosh", 2}] = "Sleef_acoshd2_u10advsimd";
+    } else if (features.vsx) {
+        retval[{"acosh", 2}] = "Sleef_acoshd2_u10vsx";
     }
 
     // atanh().
@@ -283,6 +340,10 @@ auto make_sleef_map_dbl()
         retval[{"atanh", 2}] = "Sleef_atanhd2_u10sse4";
     } else if (features.sse2) {
         retval[{"atanh", 2}] = "Sleef_atanhd2_u10sse2";
+    } else if (features.aarch64) {
+        retval[{"atanh", 2}] = "Sleef_atanhd2_u10advsimd";
+    } else if (features.vsx) {
+        retval[{"atanh", 2}] = "Sleef_atanhd2_u10vsx";
     }
 
     // erf().
@@ -298,6 +359,10 @@ auto make_sleef_map_dbl()
         retval[{"erf", 2}] = "Sleef_erfd2_u10sse4";
     } else if (features.sse2) {
         retval[{"erf", 2}] = "Sleef_erfd2_u10sse2";
+    } else if (features.aarch64) {
+        retval[{"erf", 2}] = "Sleef_erfd2_u10advsimd";
+    } else if (features.vsx) {
+        retval[{"erf", 2}] = "Sleef_erfd2_u10vsx";
     }
 
     return retval;
@@ -323,17 +388,6 @@ std::string sleef_function_name(llvm::LLVMContext &c, const std::string &f, llvm
     } else {
         return "";
     }
-}
-
-// NOTE: this function is here only to introduce a fake
-// dependency of heyoka on libsleef. This is needed
-// in certain setups that would otherwise unlink libsleef
-// from heyoka after detecting that no symbol from libsleef
-// is being used (even though we need libsleef to be linked
-// in for use in the JIT machinery).
-HEYOKA_DLL_PUBLIC double sleef_dummy_f(double x)
-{
-    return ::Sleef_sin_u10(x);
 }
 
 } // namespace heyoka::detail