+++ src/math/sigmoid.cpp
@@ -42,6 +42,7 @@
 #endif
 
 #include <heyoka/detail/llvm_helpers.hpp>
+#include <heyoka/detail/llvm_vector_type.hpp>
 #include <heyoka/detail/sleef.hpp>
 #include <heyoka/detail/string_conv.hpp>
 #include <heyoka/detail/taylor_common.hpp>
@@ -104,7 +105,7 @@ llvm::Value *sigmoid_impl::codegen_dbl(llvm_state &s, const std::vector<llvm::Va
     assert(args.size() == 1u);
     assert(args[0] != nullptr);
 
-    if (auto vec_t = llvm::dyn_cast<llvm::VectorType>(args[0]->getType())) {
+    if (auto vec_t = llvm::dyn_cast<llvm_vector_type>(args[0]->getType())) {
         const auto batch_size = boost::numeric_cast<std::uint32_t>(vec_t->getNumElements());
 
         if (const auto sfn = sleef_function_name(s.context(), "exp", vec_t->getElementType(), batch_size);